requirements.txt
selene==2.0.0rc4
pytest==7.4.0
allure-pytest==2.13.2
